feat!: extend the Storage ctx to support Mux video assets (TS-26) (#90)
This PR extends the Storage ctx to support Mux video assets as blobs.
There were some breaking rewrites of the Blob schema, so one might have
to wipe the database (`mix ecto.reset`) in order to merge it and make it
to work properly in dev environment (we haven't deployed to prod yet, so
it's OK to do so).

* feat(refactor polymorphic storage attachments): use idomatic "ecto
way" of handling polymorphic entity relationships using abstract
attachment table.

---------

Co-authored-by: Aaron Zomback <aaronzomback@gmail.com>